+++ lib/index.js
@@ -356,7 +356,7 @@ module.exports = mergeExports(fn, {
 		return require("./Stats");
 	},
 	get ManifestPlugin() {
-		return require("./stats/ManifestPlugin");
+		return require("./ManifestPlugin");
 	},
 	get Template() {
 		return require("./Template");
